% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lrv.R
\name{ar_coef}
\alias{ar_coef}
\title{Computes estimator of the AR(p) coefficients by the procedure from
Khismatullina and Vogt (2019).}
\usage{
ar_coef(data, l1, l2, correct, p)
}
\arguments{
\item{data}{Time series.}

\item{l1, l2}{Tuning parameters.}

\item{correct}{Vector of the corrections, either zero or calculated by
the function \code{\link{corrections}}.}

\item{p}{AR order of the time series.}
}
\value{
Vector of length p  of estimated AR coefficients.
}
\description{
Computes estimator of the AR(p) coefficients by the procedure from
Khismatullina and Vogt (2019).
}
\keyword{internal}
