% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_minimal_intervals.R
\name{compute_minimal_intervals}
\alias{compute_minimal_intervals}
\title{Compute the set of minimal intervals as described in Duembgen (2002)}
\usage{
compute_minimal_intervals(dataset)
}
\arguments{
\item{dataset}{Set of all intervals that have a corresponding test
statistic bigger than the respective critical value.}
}
\value{
p_t_set   Set of minimal intervals
}
\description{
The result of our multiscale test is the set of all intervals
             that have a corresponding test statistic bigger than
             the respective critical value. In order to make understandable
             statistic statements about these intervals, we need to find
             so-called minimal intervals: for a given set of intervals K,
             all intervals J such that K does not contain a proper subset
             of J are called minimal. Given K, this function computes the
             set of minimal intervals.
             Procedure is described in Duembgen (2002).
}
