% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quantiles.R
\name{compute_quantiles}
\alias{compute_quantiles}
\title{Computes quantiles of the gaussian multiscale statistics.}
\usage{
compute_quantiles(
  t_len,
  grid = NULL,
  n_ts = 1,
  sigma_vector = NULL,
  deriv_order = 0,
  sim_runs = 1000,
  probs = seq(0.5, 0.995, by = 0.005),
  epidem = FALSE
)
}
\arguments{
\item{t_len}{An integer. Sample size.}

\item{grid}{Grid of location-bandwidth points as produced by
the function \code{\link{construct_grid}}, list with
the elements 'gset', 'bws', 'gtype'. If not provided,
then the defalt grid is produced and used.}

\item{n_ts}{An integer. Number of time series analyzed. Default is 1}

\item{sigma_vector}{A numeric vector of estimated sqrt(long-run variance)
for each time series. If not given, then the default is
a vector of ones of length n_ts (1, ..., 1).}

\item{deriv_order}{An integer. Order of the derivative of the trend
that is being investigated. Default is 0.}

\item{sim_runs}{Number of simulation runs to produce quantiles.
Default is 1000.}

\item{probs}{A numeric vector of probability levels (1-alpha)
for which the quantiles are computed.
Default is probs=seq(0.5,0.995,by=0.005).}

\item{epidem}{A logical parameter. True if we are investigating
epidemiological model. Default is FALSE.}
}
\value{
quant       Matrix with 2 rows where the first row contains
                    the vector of probabilities and the second contains
                    corresponding quantile of the gaussian statistics
                    distribution.
}
\description{
Quantiles from this distribution are used to approximate
                    the quantiles for the multiscale test.
}
\examples{
compute_quantiles(100)
}
