% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_statistics}
\alias{compute_statistics}
\title{Calculates statistics in case of one time series}
\usage{
compute_statistics(t_len, data, gset, n, sigma, deriv_order)
}
\arguments{
\item{t_len}{Integer. Length of time series for analysis.}

\item{data}{Double vector of length t_len that consists the time series for analysis.}

\item{gset}{A double vector of location-bandwidth points (u_1,...,u_n,h_1,...h_n).}

\item{n}{Integer. The number of location-bandwidth points in gset.}

\item{sigma}{Double. Equal to the sqrt of the long-run variance.}

\item{deriv_order}{Integer. Order of the derivative of the trend that is investigated.
Default is 0 => we analyse whether the trend itself >< than 0.}
}
\value{
stat       A double matrix of pairwise multiscale statistics Psi_ij.

vals       A double vector of length n of kernel averages (sign included).

vals_cor   A double vector of absolute value of normalized kernel averages with
                   correction \eqn{(abs(\phi(u_1,h_1)/sigmahat) - \lambda(h_1),...,abs(\phi(u_n,h_n)/sigmahat) - \lambda(h_n))}

stat       Double. Our multiscale statistics calculated as max(vals_cor).
}
\description{
Calculates statistics in case of one time series
}
