% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lrv.R
\name{corrections}
\alias{corrections}
\title{Computes vector of correction terms for second-stage estimator
of AR parameters as described in Khismatullina, Vogt (2019).}
\usage{
corrections(coefs, var_eta, len)
}
\arguments{
\item{coefs}{Given (estimated) coefficients of the AR time series.}

\item{len}{Length of the vector of the corrections}

\item{var.eta}{Variance of the innovation term}
}
\value{
c.vec    Vector of the corrections terms of length len.
}
\description{
Computes vector of correction terms for second-stage estimator
of AR parameters as described in Khismatullina, Vogt (2019).
}
\keyword{internal}
