% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid}
\alias{covid}
\title{Number of daily new cases of infections of COVID-19 per country.}
\format{
A matrix with 99 rows and 41 columns. Each column corresponds to
one coutnry, with the name of the country (denoted by three letter) being
the name of the column.
}
\source{
\url{https://www.ecdc.europa.eu}
}
\usage{
data("covid")
}
\description{
Data on the geographic distribution of COVID-19 cases worldwide
(© ECDC [2005-2019])
}
\details{
Each entry in the dataset denotes the number of new cases of infection
per day and per country. In order to make the data comparable across
countries, we take the day of the 100th confirmed case in each country as
the starting date t = 1. This way of “normalizing” the data is
common practice (Cohen and Kupferschmidt (2020)).
}
\keyword{datasets}
