% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lrv.R
\name{emp_acf}
\alias{emp_acf}
\title{Computes autocovariances at lags 0 to p for the ell-th differences of data.}
\usage{
emp_acf(data, ell, p)
}
\arguments{
\item{data}{Time series for which we calculate autocovariances.}

\item{ell}{Order of differences used.}

\item{p}{Maximum lag for the autocovariances}
}
\value{
Vector of length (p + 1) that consists of empirical
                  autocoavariances for the corresponding lag - 1.
}
\description{
Computes autocovariances at lags 0 to p for the ell-th differences of data.
}
\keyword{internal}
