% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lrv.R
\name{estimate_lrv}
\alias{estimate_lrv}
\title{Computes estimator of the long-run variance of the error terms.}
\usage{
estimate_lrv(data, q, r_bar, p)
}
\arguments{
\item{data}{A vector of Y(1), Y(2), ... Y(T).}

\item{q, r_bar}{Integers, tuning parameters.}

\item{p}{AR order of the error terms.}
}
\value{
lrv      Estimator of the long run variance of the error terms.

ahat     Vector of length p of estimated AR coefficients.

vareta   Estimator of the variance of the innovation term
}
\description{
A difference based estimator for the coefficients and
                 long-run variance in case of a nonparametric regression
                 model \eqn{Y(t) = m(t/T) + \epsilon(t)} where the errors
                 are AR(p). The procedure was first introduced in
                 Khismatullina and Vogt (2019).
}
