% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale-package.R
\docType{package}
\name{multiscale-package}
\alias{multiscale-package}
\alias{multiscale}
\title{multiscale-package: Multiscale Inference for Nonparametric Regression(s)
with Time Series Errors}
\description{
This package performs a multiscale analysis of a nonparametric
  regression or nonparametric regressions with time series errors.

  In case of a single non-parametric regression, the multiscale method to
  test qualitative hypotheses about the nonparametric time trend \eqn{m}
  in the model \eqn{Y_t = m(t/T) + \epsilon_t} with time series errors
  \eqn{\epsilon_t} is provided. The method was first proposed in
  Khismatullina and Vogt (2019). It allows to test for shape properties
  (areas of monotonic decrease or increase) of the trend \eqn{m}.

  In case of multiple non-parametric regressions, the multiscale method
  to test qualitative hypotheses about the nonparametric time trends
  \eqn{m_i} in the model \eqn{Y_{i, t} = m_i(t/T) + \epsilon_{i, t}} with
  time series errors \eqn{\epsilon_{i, t}} is provided. The method was first
  proposed in Khismatullina and Vogt (??). It allows to test for comparison
  of the trend \eqn{m_i} and further to cluster the regressions based on
  the estiamted difference between the trends.

  All these methods require an estimator of the long-run error variance
  \eqn{\sigma^2 = \sum_{l=-\infty}^{\infty} Cov(\epsilon_0, \epsilon_l)}.
  Hence, the package also provides the difference-based
  estimator for the case that the errors belong to the class of
  \eqn{AR(\infty)} processes. The estimator was first proposed in
  Khismatullina and Vogt (2019).
}
\details{
This package performs a multiscale analysis of a nonparametric regression
or nonparametric regressions with time series errors.
}
\references{
Khismatullina M., Vogt M. Multiscale inference and long-run
            variance estimation in non-parametric regression with
            time series errors //Journal of the Royal Statistical Society:
            Series B (Statistical Methodology). - 2019.
}
\seealso{
\url{https://rss.onlinelibrary.wiley.com/doi/full/10.1111/rssb.12347}
}
