% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale_test.R
\name{multiscale_test}
\alias{multiscale_test}
\title{Carries out the multiscale test given that the values the estimatates of
long-run variance have already been computed.}
\usage{
multiscale_test(
  data,
  sigma_vec,
  n_ts = 1,
  grid = NULL,
  alpha = 0.05,
  sim_runs = 1000,
  deriv_order = 0,
  epidem = FALSE
)
}
\arguments{
\item{data}{Vector (in case of n_ts = 1) or matrix (in case of
n_ts > 1) that contains (a number of) time series
that needs to be analyzed.}

\item{sigma_vec}{Vector of estimated long-run variances.
Length must be equal to n_ts.}

\item{n_ts}{Number of time series analysed. Default is 1.}

\item{grid}{Grid of location-bandwidth points as produced by
the function \code{\link{construct_grid}}.}

\item{alpha}{Significance level. Default is 0.05.}

\item{sim_runs}{Number of simulation runs to produce quantiles.
Default is 1000.}

\item{deriv_order}{An integer. Order of the derivative of the trend
that is being investigated. Default is 0.}

\item{epidem}{Logical. If TRUE, then we are looking at an epidemic
model. Default is FALSE.}
}
\value{
quant         Quantile that wasused for testing calculated from
                      the gaussian distribution.

statistics    Value of the multiscale statistics.

test_matrix   Return in case of n_ts = 1. Matrix of test results
                      for the multiscale test defined in
                      Khismatullina and Vogt (2019).
                      test_matrix[i,j] = -1: test rejects the null for the
                               j-th location u and the i-th bandwidth h and
                               indicates a decrease in the trend
                      test_matrix[i,j] = 0:  test does not reject the null
                               for the j-th location u and the i-th
                               bandwidth h
                      test_matrix[i,j] = 1:  test rejects the null for the
                               j-th location u and the i-th bandwidth h and
                               indicates an increase in the trend
                      test_matrix[i,j] = 2:  no test is carried out at j-th
                               location u and i-th bandwidth h (because
                               the point (u, h) is excluded from the grid
                               as specified by the 'deletions' option
                               in the function \code{\link{construct_grid}})

test_matrices  Return in case of n_ts > 1. List of matrices,
                       each matrix contains test results for the pairwise
                       comparison between time series.
                       Each matrix is coded exactly as in case of n_ts = 1.

gset_with_vals Either a matrix (in case of n_ts = 1) or a list of
                       matrices (in case of n_ts > 1) that contains test
                       results together with location-bandwidth points.
}
\description{
Carries out the multiscale test given that the values the estimatates of
long-run variance have already been computed.
}
