% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sizer_map.R
\name{plot_sizer_map}
\alias{plot_sizer_map}
\title{Plots SiZer map from the test results of the multiscale testing procedure.}
\usage{
plot_sizer_map(
  u_grid,
  h_grid,
  test_results,
  plot_title = NA,
  greyscale = FALSE,
  ...
)
}
\arguments{
\item{u_grid}{Vector of location points in the unit interval \eqn{[0,1]}.}

\item{h_grid}{Vector of bandwidths from \eqn{(0,0.5)}.}

\item{test_results}{Matrix of test results created by
\code{\link{multiscale_test}}.}

\item{plot_title}{Title of the plot. Default is NA and no title is written.}

\item{greyscale}{Whether SiZer map is plotted in grey scale.
Default is FALSE.}

\item{...}{Any further options to be passed to the image function.}
}
\description{
Plots SiZer map from the test results of the multiscale testing procedure.
}
