% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_order.R
\name{select_order}
\alias{select_order}
\title{Calculates different information criterions for the number of time series
based on the long-run variance estimator (defined in
Khismatullina and Vogt (2019)) for a range of tuning parameters.}
\usage{
select_order(data, q = NULL, r = 5:15)
}
\arguments{
\item{data}{One or a number of time series in a matrix. Column names
of the matrix should be reasonable}

\item{q}{A vector of integers that consisits of different tuning
parameters to analyse. If not supplied, q is taken to be
\eqn{[2\log{T}]:([2\sqrt{T}] + 1)}.}

\item{r}{A vector of integers that consisits of different tuning
parameters r_bar to analyse. If not supplied, r = 5:15.}
}
\value{
A list with a number of elements.
orders       A vector of chosen orders of length equal to the number
             of time series.
             For each time series the order is calculated as
             \eqn{\max(which.min(fpe), ... which.min(hq))}
The rest of the elelments of the list are matrices that contain
selected orders (among 1, ..., 9) for each information criterion.
One matrix for each time series.
}
\description{
Tries to fit AR(1), ... AR(9) models for all given time series
             and calculates different information criterions (fpe, aic,
             aicc, sic, hq) for each of this fits.
}
