% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Hadley Centre Central England Temperature (HadCET) dataset,
Monthly Mean Central England Temperature (Degrees C)}
\format{
A numeric vector of length 359.
}
\source{
\url{https://www.metoffice.gov.uk/hadobs/hadcet/}
}
\usage{
data("temperature")
}
\description{
The CET dataset is the longest instrumental record of temperature
in the world. It contains the mean monthly surface air temperatures
(in degrees Celsius) from the year 1659 to the present. These monthly
temperatures are representative of a roughly triangular area of
the United Kingdom enclosed by Lancashire, London and Bristol.
Manley (1953, 1974) compiled most of the monthly series,
covering 1659 to 1973.  These data were updated to 1991 by
Parker et al (1992). It is now kept up to date by
the Climate Data Monitoring section of the Hadley Centre, Met Office.
}
\details{
Since 1974 the data have been adjusted to allow for urban warming:
currently a correction of -0.2 C is applied to mean temperatures.
CET datasets are freely available for use under Open Government License.
}
\keyword{datasets}
