% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lrv.R
\name{variance_eta}
\alias{variance_eta}
\title{Computes variance of AR(p) innovation terms eta.}
\usage{
variance_eta(data, coefs, p)
}
\arguments{
\item{data}{Time series.}

\item{coefs}{Estimated coefficients.}

\item{p}{AR order of the time series.}
}
\value{
Variance of the innovation term
}
\description{
Computes variance of AR(p) innovation terms eta.
}
\keyword{internal}
