% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_multiple_statistics}
\alias{compute_multiple_statistics}
\title{Calculates the statistics (pairwise and overall) in case of multiple time series}
\usage{
compute_multiple_statistics(
  t_len,
  n_ts,
  data,
  gset,
  n,
  sigma_vec,
  deriv_order = 0L,
  epidem = FALSE
)
}
\arguments{
\item{t_len}{Integer. Length of time series for analysis.}

\item{n_ts}{Integer. Number of time series.}

\item{data}{Double matrix t_len x n_ts. Each column consists of one time series.}

\item{gset}{A double vector of location-bandwidth points (u_1,..., u_n, h_1, ...h_n).}

\item{n}{Integer. t_lenhe number of location-bandwidth points in gset.}

\item{sigma_vec}{A double vector of length n_ts of the sqrt of long-run variances.}

\item{deriv_order}{Integer. Order of the derivative of the trend that is investigated.
Default is 0 => we analyse whether the trend itself >< than 0.}

\item{epidem}{Boolean. Equal to true in cases we are fitting epiemic model. Default is false.}
}
\value{
stat       A double matrix of pairwise multiscale statistics Psi_ij.

vals_cor   Matrix with n rows and n_ts * (n_ts - 1) / 2 columns where each column
                   contains values of the normalised Kernel averages in order to be able to perform 
                   the test on every separate interval.
}
\description{
Calculates the statistics (pairwise and overall) in case of multiple time series
}
