% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_gaussian}
\alias{simulate_gaussian}
\title{Simulates distribution of the gaussian statistics}
\usage{
simulate_gaussian(
  t_len,
  n_ts,
  sim_runs,
  gset,
  ijset,
  sigma_vec,
  deriv_order = 0L,
  epidem = FALSE
)
}
\arguments{
\item{t_len}{Integer. Length of time series for analysis.}

\item{n_ts}{Integer. Number of time series.}

\item{sim_runs}{Integer. Number of simulations needed to produce the quantiles.}

\item{gset}{A double vector of location-bandwidth points (u_1,...,u_n,h_1,...h_n).}

\item{ijset}{An integer vector of indices of the countries for the comparison
(i_1, i_2, ..., j_{n_comparisons}).}

\item{sigma_vec}{A double vector of the sqrt of long-run variances.}

\item{deriv_order}{Integer. Order of the derivative of the trend that is investigated.
Default is 0 => we analyse whether the trend itself >< than 0.}

\item{epidem}{Boolean. Equal to true in cases we are fitting epiemic model. Default is false.}
}
\value{
Phi_vec A double vector of length sim_runs that consists of the calculated Gaussian statistic for each simulation run
}
\description{
Simulates distribution of the gaussian statistics
}
